Added tractor.js and created the plantSeeds function to sort arrays and plant seeds
