+++ src/scripts/seeds/asparagus.js
@@ -1,3 +1,4 @@
+// Create the seed object
 export const createAsparagus = () => {
     const seed = {
         type: "Asparagus",
