+++ src/scripts/seeds/wheat.js
@@ -1,3 +1,4 @@
+// Create the seed object
 export const createWheat = () => {
     const seed = {
         type: "Wheat",
